+++ vision/src/autogluon/vision/detector/__init__.py
@@ -1,0 +1,1 @@
+from .detector import *
