--- vision/src/autogluon/vision/image_prediction/__init__.py
@@ -1,1 +1,0 @@
-from .image_prediction import *
