--- vision/src/autogluon/vision/object_detection/__init__.py
@@ -1,1 +1,0 @@
-from .object_detection import *
