+++ vision/src/autogluon/vision/predictor/__init__.py
@@ -1,0 +1,1 @@
+from .predictor import *
